Writing this page takes 4 days, based on the Sina News page, including search boxes, text, images, ads, and more. Doing it yourself is very rewarding.
The page I made is based on Sina.com. Including information search box, news column, picture bar, video bar, news of various topics, and added advertisements in order to improve authenticity. Finally, in accordance with the fidelity of the portal, the display of various laws and regulations in the footer is perfected.